I ate a can of sardines The sardines were in a can.
I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can. I ate a can of sardines The sardines were in a can.